requirements.txt
SQLAlchemy==1.4.5
requests==2.25.1
dnspython==2.1.0
Flask==1.1.2
flask_jwt_extended==4.1.0
flask_migrate==2.7.0
flask_restplus==0.13.0
flask_script==2.0.6
flask_sqlalchemy==2.5.1
flask_testing==0.8.1
flexpoolapi==1.2.7.post2
marshmallow==3.11.1
python-dotenv==0.17.0
timeloop==1.0.2
Werkzeug==1.0.1
